# engine/broker-server/openapi/bettermq.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BetterMQ API",
    "description": "HTTP webhook broker — publish, queues, enqueue, flow control, cron, and DLQ. Authenticate with a panel-generated `sk_local_…` Bearer token.",
    "version": "0.2.2",
    "license": {
      "name": "MIT OR Apache-2.0"
    }
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API token (`Authorization: Bearer sk_local_…`)"
      }
    },
    "schemas": {
...
      "PublishBody": {
        "type": "object",
        "required": [
          "url",
          "secret",
          "key"
        ],
        "properties": {
          "url": {
            "type": "string",
            "format": "uri"
          },
          "secret": {
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "body": {},
          "priority": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9,
            "default": 5
          },
          "flow_id": {
            "type": "string",
            "format": "uuid"
          },
          "delay": {
            "type": "integer",
            "description": "Delay in milliseconds"
          },
          "max_retries": {
            "type": "integer",
            "default": 0
          },
          "idempotency_key": {
            "type": "string"
          },
          "method": {
            "type": "string",
            "enum": [
              "GET",
              "POST",
              "PUT",
              "PATCH",
              "DELETE",
              "HEAD"
            ]
          },
          "headers": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "sign": {
            "type": "boolean"
          }
        }
      },
      "EnqueueBody": {
        "type": "object",
        "required": [
          "key"
        ],
        "properties": {
          "queue_id": {
            "type": "string",
            "format": "uuid"
          },
          "queue": {
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "body": {},
          "priority": {
            "type": "integer"
          },
          "flow_id": {
            "type": "string",
            "format": "uuid"
          },
          "delay": {
            "type": "integer"
          },
          "max_retries": {
            "type": "integer"
          },
          "idempotency_key": {
            "type": "string"
          },
          "method": {
            "type": "string"
          },
          "headers": {
            "type": "object"
          },
          "sign": {
            "type": "boolean"
          }
        }
      },
      "DeliveryRef": {
        "type": "object",
        "required": ["shard", "seq"],
        "properties": {
          "shard": { "type": "integer", "format": "uint32" },
          "seq": { "type": "integer", "format": "uint64" }
        }
      },
      "ScheduledInfo": {
        "type": "object",
        "required": ["schedule_id", "deliver_at_ms"],
        "properties": {
          "schedule_id": { "type": "string", "format": "uuid" },
          "deliver_at_ms": { "type": "integer", "format": "int64" }
        }
      },
      "AcceptResponse": {
        "type": "object",
        "required": ["queue", "duplicate"],
        "properties": {
          "message_id": { "type": "string", "format": "uuid" },
          "queue": { "type": "string" },
          "duplicate": { "type": "boolean" },
          "delivery": { "$ref": "#/components/schemas/DeliveryRef" },
          "scheduled": { "$ref": "#/components/schemas/ScheduledInfo" }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/v1/publish": {
      "post": {
        "tags": [
          "Publish"
        ],
        "operationId": "publish",
        "summary": "Publish (1:1 webhook)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Duplicate idempotency key",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AcceptResponse" }
              }
            }
          },
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AcceptResponse" }
              }
            }
          }
        }
      }
    },
...
    "/v1/enqueue": {
      "post": {
        "tags": [
          "Enqueue"
        ],
        "operationId": "enqueue",
        "summary": "Enqueue by queue name or id",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnqueueBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Duplicate idempotency key",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AcceptResponse" }
              }
            }
          },
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AcceptResponse" }
              }
            }
          }
        }
      }
    },
...
  }
}